Update README.md
This project is about a dataset of the world bank released in 2015. I have analyzed some questions based on the factors of columns. You can see a visualization of this dataset in R studio as it is written in R. It has many future scopes as based on these factors, institutions can focus on the economic growth of the country.